feat: add parameter `venv_parent`
The type of the parameter is `str` rather than `Path | str`
so the function remains self-contained
and does not require an import outside its body.

Do not make the path to `me` absolute;
there is no reason to.
We should have removed the call to `me.resolve`
instead of replacing it with `me.absolute`.